refactor: new language model folder and file structure
